created buffer.h and moved buffer related things over
